requirements/base.txt
Django==1.6.5
django-braces==1.4.0
django-model-utils==2.2
logutils==0.3.3
ipython==2.2.0
django-extensions==1.3.10
django-countries==2.1.2
django-guardian==1.2.4
djangorestframework==2.3.14
requests==2.3.0
geopy==1.1.1
